SKFP-12: Migrate app to typescript
