+++ .dockerignore
@@ -1,5 +1,3 @@
+Dockerfile
 .git
-target
 node_modules
-/dist
-.idea
