+++ .eslintrc
@@ -1,0 +1,36 @@
+{
+    "env": {
+        "commonjs": true,
+        "node": true,
+        "es6": true
+    },
+    "parser": "@typescript-eslint/parser",
+    "parserOptions": {
+        "ecmaVersion": 2020,
+        "sourceType": "module"
+    },
+    "extends": [
+        "plugin:@typescript-eslint/recommended", 
+        "plugin:prettier/recommended"
+    ],
+    "rules": {
+        "arrow-body-style": ["error", "as-needed"],
+        "complexity": ["warn", 16],
+        "template-curly-spacing": "off",
+        "indent": "off",
+        "default-case": 0,
+        "max-len": [
+            "error",
+            {
+                "code": 160,
+                "ignoreComments": true
+            }
+        ],
+        "no-console": ["error", { "allow": ["warn", "error"] }],
+        "no-unused-vars": "off",
+        "@typescript-eslint/no-unused-vars": ["warn", { "argsIgnorePattern": "^_" }],
+        "prefer-spread": "error",
+        "prettier/prettier": ["error"]
+    }
+}
+
