+++ README.md
@@ -6,16 +6,22 @@
 </p>
 
 # :factory: KF-API-ARRANGER
-This is an instantiation of the [@arranger/server](https://github.com/overture-stack/arranger/tree/develop/modules/server) application for the Kids First portal, with an integration with [Ego](https://github.com/overture-stack/ego) for authentication.
+This is an instantiation of the [@arranger/server](https://github.com/overture-stack/arranger/tree/develop/modules/server) application for the Kids First portal, with an integration with [Keycloak](https://www.keycloak.org/docs/latest/securing_apps/index.html#_nodejs_adapter) for authentication.
 
 Arranger server is an application that wraps Elasticsearch and provides a GraphQL search API for consumption by the [Kids First Portal UI](https://github.com/kids-first/kf-portal-ui).
 
 ## Development:
 
+* Execute: `npm run build` then `npm run start`
+
 ### General
+
 * Make sure that all the needed env vars point to where they should.
+
 * When adding a new env var, update the .env.example. Otherwise, an error will be thrown.
+
 * Installing dependencies: `npm install`
 
 ### Test
+
 * Execute: `npm run test`
