--- nodemon.json
@@ -1,3 +1,0 @@
-{
-  "ignore": ["dist/*"]
-}
