+++ package.json
@@ -4,10 +4,11 @@
   "description": ":factory: a data-model aware GraphQL API that sits above an Elasticsearch cluster",
   "main": "src/index.js",
   "scripts": {
-    "dev": "nodemon --exec babel-node ./src/index.js",
+    "start": "node ./dist/src/index.js",
     "test": "jest",
-    "clean": "rm -Rf ./dist",
-    "build": "npm run clean; babel ./src --out-dir ./dist --copy-files"
+    "clean": "rm -rf ./dist/* ",
+    "build": "npm run clean; tsc ",
+    "lint": "eslint . --ext .ts"
   },
   "repository": {
     "type": "git",
@@ -30,30 +31,34 @@
   "homepage": "https://github.com/kids-first/kf-api-arranger#readme",
   "dependencies": {
     "@arranger/server": "^2.9.0",
+    "@awaitjs/express": "^0.5.0",
     "@elastic/elasticsearch": "^7.9.1",
     "cors": "2.8.5",
     "dotenv-safe": "^8.2.0",
-    "express": "^4.16.2",
-    "express-async-handler": "^1.1.4",
+    "express": "^4.17.1",
     "graphql-parse-resolve-info": "4.10.0",
     "http-status-codes": "^2.1.4",
-    "keycloak-connect": "^14.0.0"
+    "keycloak-connect": "^14.0.0",
+    "lodash": "^4.17.21",
+    "regenerator-runtime": "^0.13.9"
   },
   "devDependencies": {
-    "@babel/cli": "^7.12.10",
-    "@babel/core": "^7.12.10",
-    "@babel/node": "^7.12.10",
-    "@babel/plugin-transform-runtime": "^7.12.10",
-    "@babel/preset-env": "^7.12.10",
-    "@babel/runtime": "^7.12.5",
-    "jest": "^26.6.3",
-    "nodemon": "2.0.7",
-    "prettier": "2.2.1",
-    "supertest": "^6.1.2"
-  },
-  "prettier": {
-    "printWidth": 100,
-    "trailingComma": "all",
-    "singleQuote": true
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.13",
+    "@types/jest": "^26.0.24",
+    "@types/lodash": "^4.14.171",
+    "@types/node": "^16.3.2",
+    "@types/supertest": "^2.0.11",
+    "@typescript-eslint/eslint-plugin": "^4.28.3",
+    "@typescript-eslint/parser": "^4.28.3",
+    "eslint": "^7.29.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^3.4.0",
+    "jest": "^27.0.6",
+    "prettier": "^1.19.1",
+    "supertest": "^6.1.2",
+    "ts-jest": "^27.0.4",
+    "ts-node": "^10.1.0",
+    "typescript": "^4.3.5"
   }
 }
