+++ src/env.ts
@@ -1,7 +1,6 @@
-require('dotenv-safe').config({
-  allowEmptyValues: true
-});
+import dotenv from 'dotenv';
 
+dotenv.config();
 
 export const port = process.env.PORT || 5050;
