--- src/middleware/index.js
@@ -1,2 +1,0 @@
-export { onlyAdminMutations } from './graphqlMiddleware';
-export { default as injectBodyHttpHeaders } from './injectBodyHttpHeaders';
