+++ tsconfig.json
@@ -1,0 +1,17 @@
+{
+    "compilerOptions": {
+        "resolveJsonModule": true,
+        "module": "commonjs",
+        "esModuleInterop": true,
+        "target": "es6",
+        "moduleResolution": "node",
+        "sourceMap": true,
+        "outDir": "dist",
+        "skipLibCheck": true
+    },
+    "lib": ["es2015"],
+    "exclude": ["node_modules"],
+    "include": [
+        "src/**/*"
+    ]
+}
